requirements.txt
git+https://github.com/dasylang/dasy.git
requests
pyfiglet
titanoboa